+++ src/AppNavigator.tsx
@@ -83,6 +83,8 @@ import { BrowserListByTabview } from 'screens/Home/Browser/BrowserListByTabview'
 import { MissionPoolsByTabview } from 'screens/Home/Browser/MissionPool';
 import { DeriveAccount } from 'screens/Account/DeriveAccount';
 import { useGroupYieldPosition } from 'hooks/earning';
+import { AboutSubWallet } from 'screens/Settings/AboutSubWallet';
+import { updateCurrentRoute } from 'stores/utils';
 
 interface Props {
   isAppReady: boolean;
@@ -407,6 +409,7 @@ const AppNavigator = ({ isAppReady }: Props) => {
   };
 
   const onUpdateRoute = useCallback((state: NavigationState | undefined) => {
+    updateCurrentRoute(state?.routes[state?.index]);
     setCurrentRoute(state?.routes[state?.index]);
   }, []);
 
@@ -542,6 +545,7 @@ const AppNavigator = ({ isAppReady }: Props) => {
                 <Stack.Screen name="DAppAccessDetail" component={DAppAccessDetailScreen} />
                 <Stack.Screen name="Languages" component={Languages} />
                 <Stack.Screen name="Security" component={Security} />
+                <Stack.Screen name="AboutSubWallet" component={AboutSubWallet} />
                 <Stack.Screen
                   name="ChangePassword"
                   component={ChangeMasterPassword}